pyproject.toml
[tool.poetry]
name = "flare"
version = "0.1.0"
description = "Photo Sync utility"
license = "MIT"
packages = [{ include = "flare" }]

[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.11.5"
pyexiftool = "^0.5.6"
gphoto2 = "^2.5.1"
pyusb = "^1.3.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
